Trail data layer: move intent, back-fill and git-outcome reads into their own store module
The session-intent, typed-event back-fill and per-session git-outcome
reads now live in clawmetry/trail_store.py as a mixin on LocalStore, and
the typed-column stamping for new event rows is a pure helper in
clawmetry/event_shape.py. Behaviour is unchanged; the store module stops
growing and the feature is readable in one place.
